minor improvement in CoverageUtilieis
git-svn-id: http://svn.osgeo.org/geotools/trunk@36421 e5c1c795-43da-0310-a71f-fac65c449510

+++ modules/library/coverage/src/main/java/org/geotools/resources/coverage/CoverageUtilities.java
@@ -593,9 +593,9 @@ public final class CoverageUtilities {
 		case DataBuffer.TYPE_SHORT:
 			return Short.valueOf((short)Short.MIN_VALUE);
 		case DataBuffer.TYPE_DOUBLE:
-			return Double.NaN;
+			return Double.valueOf(Double.NaN);
 		case DataBuffer.TYPE_FLOAT:
-			return Float.NaN;
+			return Float.valueOf(Float.NaN);
 		default:
 			throw new IllegalAccessError(Errors.format(ErrorKeys.ILLEGAL_ARGUMENT_$2,"dataType",dataType));
 		}
